requirements.txt
python-binance>=1.0.19
pandas>=2.2.2
numpy>=1.26.4
openai>=1.40.0
python-dotenv>=1.0.1